refactor(user): 디데이박스 이름 변경, 컴포넌트 분리
